+++ atlasdb-ete-test-utils/build.gradle
@@ -12,7 +12,6 @@ dependencies {
     api project(':atlasdb-api')
     implementation project(':atlasdb-client')
 
-    implementation 'junit:junit'
     implementation 'org.junit.jupiter:junit-jupiter'
     implementation 'org.junit.jupiter:junit-jupiter-api'
     implementation 'org.assertj:assertj-core'
